fix/#68 header 디자인 및 간격 조정
