+++ src/entities/chat/ui/ChatRoomListHeader.tsx
@@ -1,11 +1,11 @@
-import { NotificationBell } from '@/widgets/common/NotificationBell';
+import { NotificationButton } from '@/widgets/common/NotificationButton';
 
 export const ChatRoomListHeader = () => {
   return (
     <header className="safe-pt relative flex h-(--space-h-header) items-center justify-center bg-white px-4">
       <h1 className="text-xl font-semibold">채팅</h1>
-      <div className="absolute top-2 right-2">
-        <NotificationBell />
+      <div className="absolute top-4 right-4">
+        <NotificationButton />
       </div>
     </header>
   );
