+++ src/entities/feed/ui/FeedHeader.tsx
@@ -1,16 +1,13 @@
-import { Siren } from 'lucide-react';
-import { Button } from '@/shared/ui/Button';
-import { NotificationBell } from '@/widgets/common/NotificationBell';
+import { NotificationButton } from '@/widgets/common/NotificationButton';
+import { ReportButton } from '@/widgets/common/ReportButton';
 
 export const FeedHeader = () => {
   return (
     <>
       <header className="safe-pt relative flex h-(--space-h-header) items-center justify-between bg-white px-4">
-        <Button variant="ghost" size="icon">
-          <Siren className="size-6" />
-        </Button>
+        <ReportButton />
         <h1 className="text-xl font-semibold">게시판</h1>
-        <NotificationBell />
+        <NotificationButton />
       </header>
     </>
   );
