+++ src/widgets/explore/ExploreHeader.tsx
@@ -2,7 +2,8 @@
 
 import Image from 'next/image';
 import { SlidersHorizontal } from 'lucide-react';
-import { NotificationBell } from '@/widgets/common/NotificationBell';
+import { NotificationButton } from '@/widgets/common/NotificationButton';
+import { Button } from '@/shared/ui/Button';
 
 interface ExploreHeaderProps {
   /** 필터 버튼 클릭 핸들러 */
@@ -18,26 +19,18 @@ interface ExploreHeaderProps {
  */
 export const ExploreHeader = ({ onFilterClick }: ExploreHeaderProps) => {
   return (
-    <div className="sticky top-0 z-50 border-b border-gray-100 bg-white/90 backdrop-blur-sm">
-      <div className="flex h-14 items-center justify-between px-4">
-        {/* 좌측: 필터 버튼 */}
-        <button
-          onClick={onFilterClick}
-          className="flex h-10 w-10 items-center justify-center rounded-full transition-colors hover:bg-gray-100 active:bg-gray-200"
-          aria-label="필터 옵션 열기"
-          type="button"
-        >
-          <SlidersHorizontal className="size-6 text-gray-700" />
-        </button>
+    <header className="safe-pt sticky top-0 z-50 flex h-(--space-h-header) w-full items-center justify-between bg-white px-4">
+      <Button variant="ghost" size="icon" className="cursor-pointer" onClick={onFilterClick}>
+        <SlidersHorizontal className="size-6 text-gray-900" />
+      </Button>
 
-        {/* 중앙: Tokit 심볼 */}
-        <div className="flex items-center">
-          <Image src="/images/logo/tokit_symbol.svg" alt="Tokit" width={90} height={10} />
-        </div>
-
-        {/* 우측: 알림 아이콘 */}
-        <NotificationBell />
+      {/* 중앙: Tokit 심볼 */}
+      <div className="flex items-center">
+        <Image src="/images/logo/tokit_symbol.svg" alt="Tokit" width={90} height={10} />
       </div>
-    </div>
+
+      {/* 우측: 알림 아이콘 */}
+      <NotificationButton />
+    </header>
   );
 };
